requirements.txt
certifi==2020.12.5
chardet==4.0.0
get-all-tickers==1.7
idna==2.10
numpy==1.19.5
pandas==1.2.1
praw==7.1.0
prawcore==1.5.0
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
update-checker==0.18.0
urllib3==1.26.2
websocket-client==0.57.0
